Cohomological reformulation.
I've explained how to reformulate the splitting problem in terms of
cohomological vanishing, and I've explained how to eliminate the group
$G$ from the picture.

Next, I need to show how to get rid of the field $k$ and reduce the
problem to the computation of some universal cohomology groups.
Then, I should mention the problem of computational complexity,
introduce the Salvetti complex, then start talkig about the problem of
extending the computational results to infinite cases.